two-microservixes/service_ui/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
fastapi==0.103.1
googleapis-common-protos==1.63.2
grpcio==1.65.5
h11==0.14.0
idna==3.7
importlib_metadata==8.0.0
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-instrumentation-requests==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
protobuf==4.25.4
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
requests==2.31.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.23.2
wrapt==1.16.0
zipp==3.20.0
